Extract shared types to Types.swift (#470)
Extract Cart from CheckoutCompleteEvent
